requirements.txt
# === Core (all bots) ===
aiohttp>=3.9
discord-py-interactions>=5.0

# === Persona bot: Claude API ===
anthropic>=0.40

# === Persona bot: RAG (CPU-safe for Railway) ===
sentence-transformers>=2.2
chromadb>=0.4

# === Protector bot: Sentiment analysis ===
vaderSentiment>=3.3

# === Large file import (optional) ===
ijson>=3.2

# == for environment handling
python-dotenv>=1.0

# === Vision / CLIP (LOCAL ONLY - comment out for Railway) ===
# clip-interrogator>=0.6
# Pillow>=10.0
# transformers>=4.30
